fix warnings
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@9784 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ framework/pcore/src/DataStoreStreamer.cc
@@ -60,7 +60,7 @@ DataStoreStreamer::DataStoreStreamer(int complevel, int maxthread) : m_compressi
   m_threadin(0)
   //, m_threadout(0)
 {
-  if (m_maxthread > c_maxThreads) {
+  if ((unsigned int)m_maxthread > c_maxThreads) {
     B2FATAL("DataStoreStreamer : Too many threads " << m_maxthread);
     m_maxthread = c_maxThreads;
   }
